fonctionnel + début nv thread resources

+++ src/main/java/com/majeur/projet/apiCommunication/FireObject.java
@@ -3,8 +3,8 @@ package com.majeur.projet.apiCommunication;
 public class FireObject {
     private int id;
     private int intensity;
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
     private String type;
 
     public FireObject(){
@@ -34,19 +34,19 @@ public class FireObject {
         this.range = range;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
